Plot classified images

plots a classified raster using tmap.

Usage

# S3 method for class 'class_cube'
plot(
  x,
  y,
  ...,
  tile = x[["tile"]][[1L]],
  roi = NULL,
  legend = NULL,
  palette = "Spectral",
  scale = 1,
  max_cog_size = 1024L,
  legend_position = "outside"
)

Arguments

x

Object of class "class_cube".

y

Ignored.

...

Further specifications for plot.

tile

Tile to be plotted.

roi

Spatial extent to plot in WGS 84 - named vector with either (lon_min, lon_max, lat_min, lat_max) or (xmin, xmax, ymin, ymax)

legend

Named vector that associates labels to colors.

palette

Alternative RColorBrewer palette

scale

Relative scale (0.4 to 1.0) of plot text

max_cog_size

Maximum size of COG overviews (lines or columns)

legend_position

Where to place the legend (default = "outside")

Value

A color map, where each pixel has the color associated to a label, as defined by the legend parameter.

Note

The following optional parameters are available to allow for detailed control over the plot output:

graticules_labels_size: size of coord labels (default = 0.8)

legend_title_size: relative size of legend title (default = 1.0)

legend_text_size: relative size of legend text (default = 1.0)

legend_bg_color: color of legend background (default = "white")

legend_bg_alpha: legend opacity (default = 0.5)

Author

Gilberto Camara, gilberto.camara@inpe.br

Examples

if (sits_run_examples()) {
    # create a random forest model
    rfor_model <- sits_train(samples_modis_ndvi, sits_rfor())
    # create a data cube from local files
    data_dir <- system.file("extdata/raster/mod13q1", package = "sits")
    cube <- sits_cube(
        source = "BDC",
        collection = "MOD13Q1-6.1",
        data_dir = data_dir
    )
    # classify a data cube
    probs_cube <- sits_classify(
        data = cube, ml_model = rfor_model, output_dir = tempdir()
    )
    # label cube with the most likely class
    label_cube <- sits_label_classification(
        probs_cube,
        output_dir = tempdir()
    )
    # plot the resulting classified image
    plot(label_cube)
}