Plot a SOM map

plots a SOM map generated by "sits_som_map". The plot function produces different plots based on the input data. If type is "codes", plots the vector weight for in each neuron. If type is "mapping", shows where samples are mapped.

Usage

# S3 method for class 'som_map'
plot(x, y, ..., type = "codes", legend = NULL, band = NULL)

Arguments

x

Object of class "som_map".

y

Ignored.

...

Further specifications for plot.

type

Type of plot: "codes" for neuron weight (time series) and "mapping" for the number of samples allocated in a neuron.

legend

Legend with colors to be plotted

band

What band will be plotted (character)

Value

Called for side effects.

Author

Gilberto Camara, gilberto.camara@inpe.br

Examples

if (sits_run_examples()) {
    # create a SOM map
    som_map <- sits_som_map(samples_modis_ndvi)
    # plot the SOM map
    plot(som_map)
}