Export a a full sits tibble to the CSV format

Converts metadata and data from a sits tibble to a CSV file. The CSV file will not contain the actual time series. Its columns will be the same as those of a CSV file used to retrieve data from ground information ("latitude", "longitude", "start_date", "end_date", "cube", "label"), plus the all the time series for each data

sits_timeseries_to_csv(data, file = NULL)

Arguments

data

Time series (tibble of class "sits").

file

Full path of the exported CSV file (valid file name with extension ".csv").

Value

Return data.frame with CSV columns (optional)

Author

Gilberto Camara, gilberto.camara@inpe.br

Examples

csv_file <- paste0(tempdir(), "/cerrado_2classes_ts.csv")
sits_timeseries_to_csv(cerrado_2classes, file = csv_file)
#> # A tibble: 746 × 52
#>       id longitude latitude start_date end_date   label  NDVI1 NDVI2 NDVI3 NDVI4
#>    <int>     <dbl>    <dbl> <date>     <date>     <chr>  <dbl> <dbl> <dbl> <dbl>
#>  1     1     -54.2    -14.0 2000-09-13 2001-08-29 Cerra… 0.395 0.630 0.548 0.688
#>  2     2     -54.2    -14.0 2001-09-14 2002-08-29 Cerra… 0.406 0.532 0.586 0.468
#>  3     3     -54.2    -14.0 2002-09-14 2003-08-29 Cerra… 0.249 0.368 0.490 0.504
#>  4     4     -54.2    -14.0 2003-09-14 2004-08-28 Cerra… 0.334 0.376 0.658 0.644
#>  5     5     -54.2    -14.0 2004-09-13 2005-08-29 Cerra… 0.322 0.384 0.505 0.484
#>  6     6     -54.2    -14.0 2005-09-14 2006-08-29 Cerra… 0.363 0.431 0.604 0.717
#>  7     7     -54.2    -14.0 2006-09-14 2007-08-29 Cerra… 0.403 0.485 0.580 0.638
#>  8     8     -54.2    -14.0 2007-09-14 2008-08-28 Cerra… 0.366 0.353 0.413 0.501
#>  9     9     -54.2    -14.0 2008-09-13 2009-08-29 Cerra… 0.428 0.518 0.499 0.666
#> 10    10     -54.2    -14.0 2009-09-14 2010-08-29 Cerra… 0.545 0.493 0.643 0.675
#> # ℹ 736 more rows
#> # ℹ 42 more variables: NDVI5 <dbl>, NDVI6 <dbl>, NDVI7 <dbl>, NDVI8 <dbl>,
#> #   NDVI9 <dbl>, NDVI10 <dbl>, NDVI11 <dbl>, NDVI12 <dbl>, NDVI13 <dbl>,
#> #   NDVI14 <dbl>, NDVI15 <dbl>, NDVI16 <dbl>, NDVI17 <dbl>, NDVI18 <dbl>,
#> #   NDVI19 <dbl>, NDVI20 <dbl>, NDVI21 <dbl>, NDVI22 <dbl>, NDVI23 <dbl>,
#> #   EVI1 <dbl>, EVI2 <dbl>, EVI3 <dbl>, EVI4 <dbl>, EVI5 <dbl>, EVI6 <dbl>,
#> #   EVI7 <dbl>, EVI8 <dbl>, EVI9 <dbl>, EVI10 <dbl>, EVI11 <dbl>, …