Plot time series and data cubes

This is a generic function. Parameters depend on the specific type of input. See each function description for the required parameters.

sits tibble: see plot.sits

patterns: see plot.patterns

classified time series: see plot.predicted

raster cube: see plot.raster_cube

SAR cube: see plot.sar_cube

DEM cube: see plot.dem_cube

vector cube: see plot.vector_cube

classification probabilities: see plot.probs_cube

classification uncertainty: see plot.uncertainty_cube

uncertainty of vector cubes: see plot.uncertainty_vector_cube

classified cube: see plot.class_cube

classified vector cube: see plot.class_vector_cube

dendrogram cluster: see plot.sits_cluster

SOM map: see plot.som_map

SOM evaluate cluster: see plot.som_evaluate_cluster

geo-distances: see plot.geo_distances

random forest model: see plot.rfor_model

xgboost model: see plot.xgb_model

torch ML model: see plot.torch_model

Plots the time series to be used for classification

Usage

# S3 method for class 'sits'
plot(x, y, ..., together = TRUE)

Arguments

x

Object of class "sits".

y

Ignored.

...

Further specifications for plot.

together

A logical value indicating whether the samples should be plotted together.

Value

A series of plot objects produced by ggplot2 showing all time series associated to each combination of band and label, and including the median, and first and third quartile ranges.

Author

Gilberto Camara, gilberto.camara@inpe.br

Examples

if (sits_run_examples()) {
    # plot sets of time series
    plot(cerrado_2classes)
}