Plot a SOM map

plots a SOM map generated by "sits_som_map". The plot function produces different plots based on the input data. If type is "codes", plots the vector weight for in each neuron. If type is "mapping", shows where samples are mapped.

# S3 method for class 'som_map'
plot(x, y, ..., type = "codes", band = 1)

Arguments

x

Object of class "som_map".

y

Ignored.

...

Further specifications for plot.

type

Type of plot: "codes" for neuron weight (time series) and "mapping" for the number of samples allocated in a neuron.

band

What band will be plotted.

Value

Called for side effects.

Note

Please refer to the sits documentation available in <https://e-sensing.github.io/sitsbook/> for detailed examples.

Author

Gilberto Camara, gilberto.camara@inpe.br

Examples

if (sits_run_examples()) {
    # create a SOM map
    som_map <- sits_som_map(samples_modis_ndvi)
    # plot the SOM map
    plot(som_map)
}