sits

Satellite Image Time Series Analysis for Earth Observation Data Cubes

Note

The main sits classification workflow has the following steps:

sits_cube: selects a ARD image collection from a cloud provider.

sits_cube_copy: copies an ARD image collection from a cloud provider to a local directory for faster processing.

sits_regularize: create a regular data cube from an ARD image collection.

sits_apply: create new indices by combining bands of a regular data cube (optional).

sits_get_data: extract time series from a regular data cube based on user-provided labelled samples.

sits_train: train a machine learning model based on image time series.

sits_classify: classify a data cube using a machine learning model and obtain a probability cube.

sits_smooth: post-process a probability cube using a spatial smoother to remove outliers and increase spatial consistency.

sits_label_classification: produce a classified map by selecting the label with the highest probability from a smoothed cube.

Purpose

The SITS package provides a set of tools for analysis, visualization and classification of satellite image time series. It includes methods for filtering, clustering, classification, and post-processing.

See also

Author

Maintainer: Gilberto Camara gilberto.camara.inpe@gmail.com [thesis advisor]

Authors:

Rolf Simoes rolfsimoes@gmail.com

Felipe Souza felipe.carvalho@inpe.br

Felipe Carlos efelipecarlos@gmail.com

Other contributors:

Lorena Santos lorena.santos@inpe.br [contributor]

Charlotte Pelletier charlotte.pelletier@univ-ubs.fr [contributor]

Estefania Pizarro eapizarroa@ine.gob.cl [contributor]

Karine Ferreira karine.ferreira@inpe.br [contributor, thesis advisor]

Alber Sanchez alber.ipia@inpe.br [contributor]

Alexandre Assuncao alexcarssuncao@gmail.com [contributor]

Daniel Falbel dfalbel@gmail.com [contributor]

Gilberto Queiroz gilberto.queiroz@inpe.br [contributor]

Johannes Reiche johannes.reiche@wur.nl [contributor]

Pedro Andrade pedro.andrade@inpe.br [contributor]

Pedro Brito pedro_brito1997@hotmail.com [contributor]

Renato Assuncao assuncaoest@gmail.com [contributor]

Ricardo Cartaxo rcartaxoms@gmail.com [contributor]